refactor: move PushdownChecker/PushdownColumns into projection_read_plan
Makes the imports between row_filter and projection_read_plan
one-directional (row_filter -> projection_read_plan) instead of
circular, and removes a duplicated comment in
resolve_struct_field_leaves. Addresses review feedback.
